component cleanup and logout toast
